fixed aspect ratio bug

+++ programming_along/util/Camera.cpp
@@ -196,7 +196,7 @@ CameraOptix Camera::GetOptixCamera() const
 	cam.Position = Eye;
 	cam.LookingDirection = normalize(At - Eye);
 	cam.CosFovy = CosFovy;
-	const float aspectRatio = Width / Height;
+	const float aspectRatio = Width / (float)Height;
 	cam.Horizontal = CosFovy * aspectRatio * normalize(cross(cam.LookingDirection, Up));
 	cam.Vertical = CosFovy * normalize(cross(cam.Horizontal, cam.LookingDirection));
